pyproject.toml
[project]
name = "housing-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright>=0.7.0",
]
